backend/requirements.txt
azure-ai-textanalytics==5.0.0
azure-common==1.1.26
azure-core==1.12.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
idna==2.10
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
msrest==0.6.21
oauthlib==3.1.0
Pillow==8.1.2
pytesseract==0.3.7
python-dotenv==0.15.0
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
tika==1.24
urllib3==1.26.4
Werkzeug==1.0.1
